Look for new wellknown entry to search for the titleOrdered listing.
As we are now using the DirentLookup in the constructor, we may face
some error when reading broken zim files.

Those errors are already catch with ASSERT, but as is it is throw in the
constructor, unit tests of validate function fail because error is not
the same as expected. So we remplace ASSERT by a ZimFileFormatError
with the same message than the one raised by the validate function.