Correctly set the offset of the base titlePtrList.
We have no guaranty that the titlePtrList is the last blob added to the
cluster.
But the offset of the data in the cluster depends of the number of
blob in the cluster (because we add the offset at the beginning of the
cluster).

So we have to calculate the real offset at the end.
- `zim::writer::Cluster::getBlobOffset` return the offset of the blob
relative to the data start, not the beginning of the cluster.
- The blobOffsets vector is clear once we have written the cluster, so
  we must keep the count in a variable.