Do not use the url's ordered namespace range to do title search.
`getNamespaceBeginOffset` and `getNamespaceEndOffset` return offset
in the url order.

When we had the title list containing all the entries and being sorted
by `ns/title`. The offsets were the same.

But now we may have title list containing less entries. So the offsets
are not good.

The simpler is to do the search on all the title index and not narrow
down the search on the namespace.

On old zim file it is a performance lost but on new zim file it is
even as entries are all in the same namespace.